refactor: extract navbar into dedicated Navbar component
