+++ web/src/app/page.tsx
@@ -7,7 +7,7 @@ import "./globals.css";
 import { type RefObject, useCallback, useEffect, useRef, useState, useSyncExternalStore } from "react";
 import HeroAndDownload from "../components/HeroAndDownload";
 import ScrollToTop from "../components/ScrollToTop";
-import { ModeToggle } from "@/components/mode-toggle";
+import Navbar from "../components/Navbar";
 import { PageWrapper, Section } from "../components/layout";
 
 import { withBasePath } from "@/lib/basePath";
@@ -83,11 +83,7 @@ const subscribeToCursorGlow = (callback: () => void) => {
   };
 };
 
-const TRACKED_SECTION_IDS = ["screenshots", "features", "programs", "contact"];
-
 export default function Home() {
-  const [scrolled, setScrolled] = useState(false);
-  const [mobileMenuOpen, setMobileMenuOpen] = useState(false);
   const [comingSoonModal, setComingSoonModal] = useState(false);
   const [appleModal, setAppleModal] = useState(false);
   const [testerEmail, setTesterEmail] = useState("");
@@ -96,7 +92,6 @@ export default function Home() {
   const [currentScreenshot, setCurrentScreenshot] = useState(0);
   const [userSliderOpen, setUserSliderOpen] = useState(true);
   const [adminSliderOpen, setAdminSliderOpen] = useState(false);
-  const [activeSection, setActiveSection] = useState<string>("");
   const [featuresLoading, setFeaturesLoading] = useState(true);
 
   // ── DNA helix cursor ──
@@ -223,14 +218,6 @@ export default function Home() {
       dnaCanvasRef.current = null;
     };
   }, []);
-
-  // ── Scroll header ──
-  useEffect(() => {
-    const handleScroll = () => setScrolled(window.scrollY > 50);
-    window.addEventListener("scroll", handleScroll);
-    return () => window.removeEventListener("scroll", handleScroll);
-  }, []);
-
   // ── Scroll reveal ──
   useEffect(() => {
     if (window.matchMedia("(prefers-reduced-motion: reduce)").matches) {
@@ -255,45 +242,7 @@ export default function Home() {
     return () => clearTimeout(timer);
   }, []);
 
-  // Active section tracking via IntersectionObserver
-  useEffect(() => {
-    const visibleSections = new Set<string>();
-
-    const observer = new IntersectionObserver(
-      (entries) => {
-        entries.forEach((entry) => {
-          if (entry.isIntersecting) {
-            visibleSections.add(entry.target.id);
-          } else {
-            visibleSections.delete(entry.target.id);
-          }
-        });
-
-        if (visibleSections.size > 0) {
-          const topSection = TRACKED_SECTION_IDS
-            .filter((id) => visibleSections.has(id))
-            .map((id) => ({
-              id,
-              top: document.getElementById(id)?.getBoundingClientRect().top ?? Infinity,
-            }))
-            .sort((a, b) => Math.abs(a.top) - Math.abs(b.top))[0];
-
-          if (topSection) setActiveSection(topSection.id);
-        } else {
-          setActiveSection("");
-        }
-      },
-      { threshold: 0.3 }
-    );
-
-    TRACKED_SECTION_IDS.forEach((id) => {
-      const el = document.getElementById(id);
-      if (el) observer.observe(el);
-    });
-
-    return () => observer.disconnect();
-  }, []);
-
+  // Screenshot keyboard nav
   // Screenshot keyboard nav
   const navigateScreenshot = useCallback((dir: number) => {
     setCurrentScreenshot((prev) => {
@@ -402,7 +351,7 @@ const moveSlider = (ref: RefObject<HTMLDivElement | null>, dir: number) => {
     return () => clearInterval(interval);
   }, [adminSliderOpen]);
 
-  // ── TestFlight invite ──
+// ── TestFlight invite ──
   const sendTesterEmail = async () => {
     const trimmedTesterEmail = testerEmail.trim();
     if (!isValidEmail(trimmedTesterEmail)) {
@@ -429,7 +378,7 @@ const moveSlider = (ref: RefObject<HTMLDivElement | null>, dir: number) => {
 
   // ── Contact form submit → uhsocial.in API ──
   // Backend route needed: POST /api/contact/send on NEXT_PUBLIC_API_BASE_URL
-  // See /contact_newsletter_guide.md for the Express route implementation
+ // See /contact_newsletter_guide.md for the Express route implementation
   const handleContactSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
     const trimmedName = contactName.trim();
@@ -500,102 +449,9 @@ const moveSlider = (ref: RefObject<HTMLDivElement | null>, dir: number) => {
   };
 
   const selectedScreenshot = allScreenshots[currentScreenshot] ?? allScreenshots[0];
-
   return (
     <>
-
-      {/* ── Header ── */}
-      <header className={`header${scrolled ? " scrolled" : ""}`} id="header">
-        <PageWrapper as="div" className="nav">
-          <Link href={withBasePath("/")} className="logo">
-            <div className="logo-icon">
-              <Image
-                src="https://raw.githubusercontent.com/SB2318/UltimateHealth/refs/heads/main/frontend/src/assets/images/adaptive-icon.png"
-                alt="UltimateHealth Logo" width={48} height={48}
-                priority
-              />
-            </div>
-            Ultimate-Health
-          </Link>
-
-          <ul className="nav-links">
-            <li>
-              <a
-                href="#features"
-                className={`nav-link-item${activeSection === "features" ? " active" : ""}`}
-                aria-current={activeSection === "features" ? "location" : undefined}
-              >
-                <i className="fas fa-star nav-item-icon" aria-hidden="true"></i>
-                <span className="nav-item-text">Platform Highlights</span>
-              </a>
-            </li>
-            <li>
-              <a
-                href="#screenshots"
-                className={`nav-link-item${activeSection === "screenshots" ? " active" : ""}`}
-                aria-current={activeSection === "screenshots" ? "location" : undefined}
-              >
-                <i className="fas fa-image nav-item-icon" aria-hidden="true"></i>
-                <span className="nav-item-text">Screenshots</span>
-              </a>
-            </li>
-            <li>
-              <a
-                href="#programs"
-                className={`nav-link-item${activeSection === "programs" ? " active" : ""}`}
-                aria-current={activeSection === "programs" ? "location" : undefined}
-              >
-                <i className="fas fa-code-branch nav-item-icon" aria-hidden="true"></i>
-                <span className="nav-item-text">Community Programs</span>
-              </a>
-            </li>
-            <li>
-              <Link href={withBasePath("/articles")} className="nav-link-item">
-                <i className="fas fa-file-lines nav-item-icon" aria-hidden="true"></i>
-                <span className="nav-item-text">Read Articles</span>
-              </Link>
-            </li>
-            <li>
-              <Link href={withBasePath("/medical-glossary")} className="nav-link-item">
-                <i className="fas fa-book-medical nav-item-icon" aria-hidden="true"></i>
-                <span className="nav-item-text">Medical Glossary</span>
-              </Link>
-            </li>
-            <li>
-              <Link href={withBasePath("/contribute")} className="nav-link-item">
-                <i className="fas fa-users nav-item-icon" aria-hidden="true"></i>
-                <span className="nav-item-text">Join Us to Contribute</span>
-              </Link>
-            </li>
-            <li className="nav-toggle-item">
-              <ModeToggle />
-            </li>
-            <li style={{ display: "flex", alignItems: "center" }}>
-              <a href="#downloads" className="nav-btn-sm">
-                <i className="fas fa-user" aria-hidden="true"></i>
-                <span>Login / Register</span>
-              </a>
-            </li>
-          </ul>
-
-          <button className="mobile-menu-toggle" onClick={() => setMobileMenuOpen((o) => !o)} aria-label={mobileMenuOpen ? "Close navigation menu" : "Open navigation menu"} aria-expanded={mobileMenuOpen}>
-            <i className={`fas fa-${mobileMenuOpen ? "times" : "bars"}`}></i>
-          </button>
-        </PageWrapper>
-
-        <nav className={`mobile-nav${mobileMenuOpen ? " open" : ""}`}>
-          <a href="#screenshots" onClick={() => setMobileMenuOpen(false)}>Screenshots</a>
-          <a href="#features" onClick={() => setMobileMenuOpen(false)}>Platform Highlights</a>
-          <a href="#programs" onClick={() => setMobileMenuOpen(false)}>Community Programs</a>
-          <Link href={withBasePath("/articles")} onClick={() => setMobileMenuOpen(false)}>Read Articles</Link>
-          <Link href={withBasePath("/medical-glossary")} onClick={() => setMobileMenuOpen(false)}>Medical Glossary</Link>
-          <Link href={withBasePath("/contribute")} onClick={() => setMobileMenuOpen(false)}>Join Us to Contribute</Link>
-          <a href="#downloads" onClick={() => setMobileMenuOpen(false)}>Login / Register</a>
-          <div style={{ display: "flex", justifyContent: "center", padding: "8px 0" }}>
-            <ModeToggle />
-          </div>
-        </nav>
-      </header>
+<Navbar />
 
       {/* ── Hero ── */}
       <HeroAndDownload
